requirements.txt
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
xgboost>=2.0.0
plotly>=5.18.0
streamlit>=1.30.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyarrow>=14.0.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.11.0
kaggle>=1.6.0
openmeteo-requests>=1.2.0
requests-cache>=1.1.0
retry-requests>=2.0.0
